Move benefit claim call to existing service
